requirements.txt
discord.py
motor
PyYAML
python-dotenv
PyMongo
ffmpeg-python

yt-dlp